feat: queries에서 api 따로 분류
